Set Rx Connect Speed as mandatory field
xl2tpd currently sets the M-bit to 0 as RFC2661 imposes.
Unfortunately Watchguard expects the M-bit to be 1 when this attribute is present.
It's unclear to me whether completely excluding Rx Connect Speed (as 210eb9904f85eccdb3f16f6e276d7fd78d4dbd10 suggests) is a good idea or not
since that seems to imply the connection is symmertic according to RFC2661.
Also I don't currently use other L2TP VPNs, this is fine for my usecase.